requirements.txt
tensorflow
keras
numpy
pillow
matplotlib
flet
